+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format of this file is based on [Keep a Changelog] and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- Floats were not parsed correctly ([#1])
+
 ## 0.1.0 (2017-05-12)
 
 ### Added
@@ -12,4 +18,8 @@ The format of this file is based on [Keep a Changelog] and this project adheres 
 
 [Keep a Changelog]: http://keepachangelog.com/
 [Semantic Versioning]: http://semver.org/
+
+[Unreleased]: https://github.com/torfsen/barely_json/compare/v0.1.0...master
+
+[#1]: https://github.com/torfsen/barely_json/issues/1