+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The format of this file is based on [Keep a Changelog] and this project adheres to [Semantic Versioning].
 
-## [Unreleased]
+## [0.1.1] (2017-05-12)
 
 ### Fixed
 
@@ -20,7 +20,7 @@ The format of this file is based on [Keep a Changelog] and this project adheres 
 [Keep a Changelog]: http://keepachangelog.com/
 [Semantic Versioning]: http://semver.org/
 
-[Unreleased]: https://github.com/torfsen/barely_json/compare/v0.1.0...master
+[0.1.1]: https://github.com/torfsen/barely_json/compare/v0.1.0...v0.1.1
 
 [#1]: https://github.com/torfsen/barely_json/issues/1
 [#2]: https://github.com/torfsen/barely_json/issues/2